+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-24
+- `18:11` -- **Updated** `apps/desktopProbe/src/server/rendererIpcApi.ts`
 - `18:10` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
 - `18:09` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
 - `17:38` -- **Updated** `apps/desktopProbe/src/index.ts`